Testing integration, path hackery

+++ shole/managers.py
@@ -7,11 +7,7 @@ from multiprocessing.context import TimeoutError as TimesUpPencilsDown
 from multiprocessing import Queue as MultiQueue
 from queue import Queue
 from queue import Empty as EmptyQueue
-
-
-KILL = "CLOSE"
-DONE = "DONE"
-CZEC = "CHECK"
+from constants import KILL, DONE, CZEC
 
 
 def clear_and_close_queues(*queues):
@@ -93,6 +89,7 @@ class ProcessHost(object):
         self.kill_signal = kill_signal  # unintended process opening / closing, especially with reuse.
         self.finished_signal = finished_signal
         self.check_signal = check_signal
+        self.is_running = False
         self._continue_running = run_process
         self._current_processor = None
         assert (self.kill_signal
@@ -102,9 +99,9 @@ class ProcessHost(object):
         self.process_end_signals = {self.kill_signal, self.finished_signal, self.check_signal}
         if process_target is not None and run_process:
             self.make_single_process_handler(process_target,
-                                             process_args=process_args,
+                                             *process_args,
                                              host_to_process_signals=host_to_process_signals,
-                                             process_kwarg_dict=process_kwarg_dict)
+                                             **process_kwarg_dict)
 
     def make_single_process_handler(self, process_target, *process_args,
                                     host_to_process_signals=None,
@@ -120,19 +117,33 @@ class ProcessHost(object):
         :rtype: None
         :return: None
         """
+        assert not self.is_running, ("Please create a new SingleProcessHandler to start another process while this one "
+                                     "is still running.")
         _handler_to_process_queue = MultiQueue() if host_to_process_signals else None
         self._continue_running = True
+        self.is_running = True
         self._current_processor = SingleProcessHandler(process_target,
                                                        self._to_handler_queue,
-                                                       self._to_host_queue, process_args=process_args,
+                                                       self._to_host_queue, *process_args,
                                                        handler_to_process_queue=_handler_to_process_queue,
                                                        finished_signal=self.finished_signal,
                                                        kill_signal=self.kill_signal,
                                                        check_signal=self.check_signal,
                                                        host_to_process_signals=host_to_process_signals,
-                                                       process_kwarg_dict=process_kwarg_dict).start()
+                                                       **process_kwarg_dict).start()
         self.root.after(self.message_check_rate, self.check_message)
         self.root.after(self.running_check_delay, self.check_running)
+
+    def send_signal(self, signal):
+        """
+        Send signal to other process.
+
+        :Parameters:
+            :param signal: pickle-able object sent to subprocess.
+        :rtype: None
+        :return: None
+        """
+        self._to_handler_queue.put(signal)
 
     def check_message(self, *, message_callback=None):
         """
@@ -158,9 +169,10 @@ class ProcessHost(object):
                 except EmptyQueue:
                     pass
                 else:
-                    if msg in self.process_end_signals:
-                        say_check_one_more_time = False
-                        self.kill_process(need_to_signal=False)
+                    if isinstance(msg, str):
+                        if msg in self.process_end_signals:
+                            say_check_one_more_time = False
+                            self.kill_process(need_to_signal=False)
                     message_callback(msg)
                 finally:
                     if say_check_one_more_time:
@@ -189,6 +201,7 @@ class ProcessHost(object):
             self._current_processor.join()
         clear_queues(self._to_host_queue, self._to_handler_queue)
         self._current_processor = None
+        self.is_running = False
 
     def check_running(self):
         """
@@ -202,6 +215,76 @@ class ProcessHost(object):
                 and self._current_processor.is_alive()):
             self._to_handler_queue.put(self.check_signal)
             self.root.after(self.running_check_delay, self.check_running)
+
+
+class GreedyProcessHost(ProcessHost):
+    """
+    Multiprocessing/threading object which can be accessed directly within libraries like Tkinter.
+
+    Send a target function to an instance of this class during __init__ or make_single_process_handler, and it will
+    ensure the process completes, and pass any queue return messages to the function provided in message_callback
+    during __init__.
+    """
+    def __init__(self, *args, **kwargs):
+        """Create private inter-process communication for a potentially newly started process.
+
+        :Parameters:
+            :param Tkinter.Tk root: root / object with .after(delay, callback) used for scheduling.
+            :param function message_callback: function / method used to process a message if received.
+            :param function process_target: function / method to be run asynchronously.
+            :param process_args: positional arguments to be passed to process_target.
+            :param int message_check_delay: how often message checks are scheduled using root.
+            :param int running_check_delay: how often checks on whether the subprocess is running are run.
+            :param bool run_process: determines whether to run the process_target immediately after __init__.
+            :param dict host_to_process_signals: messages for the asynchronous process which may be sent to the handler.
+            :param str finished_signal: message to be used to indicate that the asynchronous process finished.
+            :param str kill_signal: message to be used to finish the asynchronous process early.
+            :param str check_signal: message to be used to check if the asynchronous process is still alive.
+            :param process_kwarg_dict: dictionary to be passed to process_target.
+        :rtype: None
+        :return: None
+        """
+        super(GreedyProcessHost, self).__init__(*args, **kwargs)
+
+    def check_message(self, *, message_callback=None):
+        """
+        Initiate callbacks from inter-process communication. Overwrites the original check_message method in order
+        to only pull the most recent queue item.
+
+        :Parameters:
+            :param function message_callback: function / method used to process a message if received. Currently
+                determines if check_message is subsequently called as well. If intending to check for a message
+                independent of the auto-check (and not intending to start another check_message thread chain)
+                pass self.message_callback or any other function as a parameter to prevent the check_message chain.
+        :rtype: None
+        :return: None
+        """
+        if message_callback is None:
+            say_check_one_more_time = self._continue_running
+            message_callback = self.message_callback
+        else:
+            say_check_one_more_time = False
+        try:
+            if not self._to_host_queue.empty():
+                msg = None
+                try:
+                    while not self._to_host_queue.empty():
+                        msg = self._to_host_queue.get_nowait()
+                except EmptyQueue:
+                    pass
+                else:
+                    if isinstance(msg, str):
+                        if msg in self.process_end_signals:
+                            say_check_one_more_time = False
+                            self.kill_process(need_to_signal=False)
+                    message_callback(msg)
+                finally:
+                    if say_check_one_more_time:
+                        self.root.after(self.message_check_rate, self.check_message)
+            elif say_check_one_more_time:
+                self.root.after(self.message_check_rate, self.check_message)
+        except AttributeError:
+            self.kill_process()
 
 
 class SingleProcessHandler(Thread):
@@ -265,6 +348,15 @@ class SingleProcessHandler(Thread):
             self.process_args = process_args
         elif process_kwarg_dict:
             self.process_args = (process_kwarg_dict,)
+        if self.handler_to_process_queue:
+            if self.process_args:
+                self.process_args = (self.to_handler_queue, self.handler_to_process_queue) + self.process_args
+            else:
+                self.process_args = (self.to_handler_queue, self.handler_to_process_queue)
+        elif self.process_args:
+            self.process_args = (self.to_handler_queue,) + self.process_args
+        else:
+            self.process_args = (self.to_handler_queue,)
 
     def run(self):
         """
@@ -289,16 +381,19 @@ class SingleProcessHandler(Thread):
         """
         should_run = True
         msg = self.to_handler_queue.get()
-        if msg in self.end_sigs:
-            self._kill_process()
-            self.handler_to_host_queue.put(msg)
-            should_run = False
-        elif msg == self.check_signal:
-            if not self.handled_process.is_alive():
+        if isinstance(msg, str):
+            if msg in self.end_sigs:
+                self._kill_process()
                 self.handler_to_host_queue.put(msg)
                 should_run = False
-        elif msg in self.host_to_process_signals:
-            self.handler_to_process_queue.put(msg)
+            elif msg == self.check_signal:
+                if not self.handled_process.is_alive():
+                    self.handler_to_host_queue.put(msg)
+                    should_run = False
+            elif msg in self.host_to_process_signals:
+                self.handler_to_process_queue.put(msg)
+            else:
+                self.handler_to_host_queue.put(msg)
         else:
             self.handler_to_host_queue.put(msg)
         return should_run
